M2-4969: define Banner icon SVGs as consts
