+++ src/shared/components/Banners/Banner/Banner.const.tsx
@@ -1,0 +1,10 @@
+import { Svg } from 'shared/components/Svg';
+
+const getSvg = (id: string) => <Svg id={id} width={32} height={32} />;
+
+export const BANNER_ICONS = {
+  info: getSvg('more-info-filled'),
+  success: getSvg('check-circle'),
+  warning: getSvg('exclamation-circle'),
+  error: getSvg('exclamation-octagon'),
+};
